Теперь character.jsx отрисовывает карточку персонажа. Компонент отдельной страницы персонажа переименовал в CharacterPage. В app.css просто лишнее удалил
