+++ src/App.css
@@ -25,8 +25,6 @@ body {
 }
 
 .App {
-  /* display: flex; */
-  /* justify-content: center; */
   width: 75%;
   background: rgba(1, 1, 1, 0.6);
   height: 100vh;
@@ -41,9 +39,6 @@ body {
   align-items: center;
   width: 100%;
   height: 50px;
-}
-.nav__links {
-  
 }
 
 .navbar__links {
