+++ src/pages/Characters.jsx
@@ -1,21 +1,14 @@
 import React, { useContext} from 'react'
-import { useNavigate } from 'react-router-dom'
 import { CharContext } from '../context/CharContext'
+import Character from './Character'
 
 const Characters = () => {
-  const router = useNavigate()
-  const {Chars, setChars} = useContext(CharContext)
+  
+  const {chars, setChars} = useContext(CharContext)
   return ( 
     <div className='main__characters'>
-      {Chars.map(char => ( 
-      <div key={char.id} className="card__character">
-        <div>
-          <h2>{char.name}</h2>
-          <div>Класс: {char.info.class}</div>
-          <div>Здоровье: {char.info.health}</div>
-        </div>
-        <button onClick={() => router('/character/1')}>ОПИСАНИЕ</button>
-      </div>
+      {chars.map(char => ( 
+        <Character key={char.id} char={char}/>
       ))} 
     </div>
   )
